♻️ Refactor TableNode types and integrate highlightNodesAndEdges utility
