+++ .changeset/smooth-terms-appear.md
@@ -1,0 +1,5 @@
+---
+"@liam-hq/erd-core": patch
+---
+
+♻️ Refactor TableNode types and integrate highlightNodesAndEdges utility
