+++ frontend/packages/erd-core/src/components/ERDRenderer/components/ERDContent/ERDContent.tsx
@@ -1,6 +1,7 @@
 import { selectTableLogEvent } from '@/features/gtm/utils'
 import { repositionTableLogEvent } from '@/features/gtm/utils/repositionTableLogEvent'
 import { MAX_ZOOM, MIN_ZOOM } from '@/features/reactflow/constants'
+import { highlightNodesAndEdges } from '@/features/reactflow/utils'
 import { useIsTouchDevice } from '@/hooks'
 import { useVersion } from '@/providers'
 import { useUserEditingActiveStore } from '@/stores'
@@ -22,7 +23,6 @@ import { NonRelatedTableGroupNode } from './NonRelatedTableGroupNode'
 import { RelationshipEdge } from './RelationshipEdge'
 import { Spinner } from './Spinner'
 import { TableNode } from './TableNode'
-import { highlightNodesAndEdges } from './highlightNodesAndEdges'
 import type { DisplayArea } from './types'
 import { useInitialAutoLayout } from './useInitialAutoLayout'
 import { usePopStateListener } from './usePopStateListener'
