+++ frontend/packages/erd-core/src/components/ERDRenderer/components/ERDContent/TableNode/TableNode.tsx
@@ -1,3 +1,4 @@
+import type { TableNodeType } from '@/features/reactflow/types'
 import { useUserEditingStore } from '@/stores'
 import type { Node, NodeProps } from '@xyflow/react'
 import clsx from 'clsx'
@@ -5,7 +6,6 @@ import type { FC } from 'react'
 import { TableColumnList } from './TableColumnList'
 import { TableHeader } from './TableHeader'
 import styles from './TableNode.module.css'
-import type { TableNodeType } from './type'
 
 export const isTableNode = (node: Node): node is TableNodeType =>
   node.type === 'table'
