+++ frontend/packages/erd-core/src/components/ERDRenderer/components/ERDContent/TableNode/index.ts
@@ -1,2 +1,1 @@
 export * from './TableNode'
-export * from './type'
