+++ frontend/packages/erd-core/src/features/reactflow/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './columnHandleId'
 export * from './convertDBStructureToNodes'
+export * from './highlightNodesAndEdges'
